backend/ai-services/waste-classifier/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
ultralytics>=8.0.0
pydantic-settings>=2.0.0
gdown>=4.7.1
pillow>=10.0.0
opencv-python-headless>=4.8.0
